+++ src/bridge/handlers/__init__.py
@@ -4,7 +4,7 @@ Handlers pull from bootstrap registries, run builders to compose models,
 execute pipelines, and coordinate external side effects.
 """
 
-from .create_pr_from_meta import create_pr_issues_from_meta
+from .create_pr_issues_from_meta import create_pr_issues_from_meta
 from .extract_meta_from_repo import extract_meta_from_repo
 
 __all__ = [
